MVT: fixes int overflow. Fixes https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=5955. Credit to OSS Fuzz
git-svn-id: https://svn.osgeo.org/gdal/trunk@41402 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/mvt/ogrmvtdataset.cpp
@@ -837,7 +837,7 @@ void OGRMVTLayer::GetXY(int nX, int nY, double& dfX, double& dfY)
     else
     {
         dfX = nX;
-        dfY = static_cast<int>(m_nExtent) - nY;
+        dfY = static_cast<double>(m_nExtent) - nY;
     }
 }
